+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- lock SPI device file
+
+### Fixed
+- reconfigure receiver after receiving a packet with unexpected length
+  (receiver possibly accessed by other process)
 
 ## [0.2.0] - 2020-12-07
 ### Added